# promethium/promethium-openapi.yaml
openapi: 3.0.2
info:
  title: Promethium REST API
  version: 0.1.0
  license:
    name: QC Ware Terms of Service
    url: https://www.promethium.qcware.com/terms-conditions
  contact:
    name: QC Ware Corp
    email: promethium@qcware.com
    url: https://promethium.qcware.com
servers:
  - url: https://api.promethium.qcware.com
security:
  - ApiKeyAuth: []
paths:
  /v0/workflows:
    get:
      tags:
        - Workflows
      summary: List Workflows
      description: Gets a list of workflows that match the query filters in the given order
      operationId: list_workflows
      parameters:
        - description: The type of workflow; can be one of ['TorsionScan', 'ConformerSearch', 'SinglePointCalculation', 'GeometryOptimization', 'InteractionEnergyCalculation', 'ReactionPathOptimization', 'TransitionStateOptimization', 'TransitionStateOptimizationFromEndpoints']
          required: true
          schema:
            allOf:
              - $ref: '#/components/schemas/WorkflowKind'
            description: The type of workflow; can be one of ['TorsionScan', 'ConformerSearch', 'SinglePointCalculation', 'GeometryOptimization', 'InteractionEnergyCalculation', 'ReactionPathOptimization', 'TransitionStateOptimization', 'TransitionStateOptimizationFromEndpoints']
          name: kind
          in: query
        - required: false
          schema:
            title: Search
            type: string
          name: search
          in: query
        - required: false
          schema:
            title: User Id
            type: string
          name: user_id
          in: query
        - description: The status of the workflow; can be one of ['CANCELED', 'COMPLETED', 'FAILED', 'RUNNING', 'TERMINATED', 'TIMED_OUT']
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/WorkflowStatus'
            description: The status of the workflow; can be one of ['CANCELED', 'COMPLETED', 'FAILED', 'RUNNING', 'TERMINATED', 'TIMED_OUT']
            default: []
          name: status
          in: query
        - required: false
          schema:
            title: Started Before
            type: string
            format: date-time
          name: started_before
          in: query
        - required: false
          schema:
            title: Started After
            type: string
            format: date-time
          name: started_after
          in: query
        - required: false
          schema:
            title: Stopped Before
            type: string
            format: date-time
          name: stopped_before
          in: query
        - required: false
          schema:
            title: Stopped After
            type: string
            format: date-time
          name: stopped_after
          in: query
        - description: The workflow field to order by. Can be one of ['id', 'name', 'kind', 'created_at', 'last_updated_at', 'started_at', 'stopped_at', 'status']. The default is created_at.
          required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/WorkflowOrderField'
            description: The workflow field to order by. Can be one of ['id', 'name', 'kind', 'created_at', 'last_updated_at', 'started_at', 'stopped_at', 'status']. The default is created_at.
            default: created_at
          name: order_by
          in: query
        - description: Can be one of asc and desc.
          required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderDir'
            description: Can be one of asc and desc.
            default: desc
          name: order_dir
          in: query
        - required: false
          schema:
            title: Page
            minimum: 1
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Size
            maximum: 1000
            minimum: 1
            type: integer
            default: 100
          name: size
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Workflow_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Workflows
      summary: Create Workflow
      description: Create a new workflow
      operationId: create_workflow
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              title: Request Input
              oneOf:
                - $ref: '#/components/schemas/CreateConformerSearchWorkflowRequest'
                - $ref: '#/components/schemas/CreateGeometryOptimizationWorkflowRequest'
                - $ref: '#/components/schemas/CreateInteractionEnergyCalculationWorkflowRequest'
                - $ref: '#/components/schemas/CreateReactionPathOptimizationWorkflowRequest'
                - $ref: '#/components/schemas/CreateSinglePointCalculationWorkflowRequest'
                - $ref: '#/components/schemas/CreateTorsionScanWorkflowRequest'
                - $ref: '#/components/schemas/CreateTransitionStateOptimizationWorkflowRequest'
                - $ref: '#/components/schemas/CreateTransitionStateOptimizationFromEndpointsWorkflowRequest'
              discriminator:
                propertyName: kind
                mapping:
                  ConformerSearch: '#/components/schemas/CreateConformerSearchWorkflowRequest'
                  GeometryOptimization: '#/components/schemas/CreateGeometryOptimizationWorkflowRequest'
                  InteractionEnergyCalculation: '#/components/schemas/CreateInteractionEnergyCalculationWorkflowRequest'
                  ReactionPathOptimization: '#/components/schemas/CreateReactionPathOptimizationWorkflowRequest'
                  SinglePointCalculation: '#/components/schemas/CreateSinglePointCalculationWorkflowRequest'
                  TorsionScan: '#/components/schemas/CreateTorsionScanWorkflowRequest'
                  TransitionStateOptimization: '#/components/schemas/CreateTransitionStateOptimizationWorkflowRequest'
                  TransitionStateOptimizationFromEndpoints: '#/components/schemas/CreateTransitionStateOptimizationFromEndpointsWorkflowRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v0/workflows/{workflow_id}:
    get:
      tags:
        - Workflows
      summary: Get Workflow
      description: Gets the workflow with the given id
      operationId: get_workflow
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
            format: uuid4
          name: workflow_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Workflows
      summary: Delete Workflow
      description: Deletes the workflow with the given id.
      operationId: delete_workflow
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
            format: uuid4
          name: workflow_id
          in: path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v0/workflows/{workflow_id}/results:
    get:
      tags:
        - Workflows
      summary: Get Workflow Results Numeric
      operationId: get_workflow_results_numeric
      parameters:
        - required: true
          schema:
            title: Workflow Id
            type: string
            format: uuid4
          name: workflow_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowNumericResults'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    WorkflowKind:
      title: WorkflowKind
      enum:
        - TorsionScan
        - ConformerSearch
        - SinglePointCalculation
        - GeometryOptimization
        - InteractionEnergyCalculation
        - ReactionPathOptimization
        - TransitionStateOptimization
        - TransitionStateOptimizationFromEndpoints
      type: string
      description: An enumeration.
    WorkflowStatus:
      title: WorkflowStatus
      enum:
        - CANCELED
        - COMPLETED
        - FAILED
        - RUNNING
        - TERMINATED
        - TIMED_OUT
      type: string
      description: Possible state values used in workflows
    WorkflowOrderField:
      title: WorkflowOrderField
      enum:
        - id
        - name
        - kind
        - created_at
        - last_updated_at
        - started_at
        - stopped_at
        - status
      type: string
      description: An enumeration.
    OrderDir:
      title: OrderDir
      enum:
        - asc
        - desc
      type: string
      description: An enumeration.
    GpuType:
      title: GpuType
      enum:
        - a100
        - v100
      type: string
      description: The classes of computing units supported in Promethium
    ResourceRequest:
      title: ResourceRequest
      required:
        - gpu_type
      type: object
      properties:
        gpu_type:
          allOf:
            - $ref: '#/components/schemas/GpuType'
          description: The class of GPU used to run the workflow; can be one of ['a100', 'v100']
        gpu_count:
          title: Gpu Count
          type: integer
          description: The number of GPU units used to run the workflow
          default: 1
      additionalProperties: false
    Workflow:
      title: Workflow
      required:
        - id
        - name
        - kind
        - created_at
        - last_updated_at
        - status
        - duration_seconds
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        id:
          title: Id
          type: string
          description: The unique id of the workflow
          format: uuid4
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          allOf:
            - $ref: '#/components/schemas/WorkflowKind'
          description: The type of workflow; can be one of ['TorsionScan', 'ConformerSearch', 'SinglePointCalculation', 'GeometryOptimization', 'InteractionEnergyCalculation', 'ReactionPathOptimization', 'TransitionStateOptimization', 'TransitionStateOptimizationFromEndpoints']
        created_at:
          title: Created At
          type: string
          description: The creation date of the workflow
          format: date-time
        last_updated_at:
          title: Last Updated At
          type: string
          description: The last update date of the workflow
          format: date-time
        started_at:
          title: Started At
          type: string
          description: The date when the workflow started
          format: date-time
        stopped_at:
          title: Stopped At
          type: string
          description: The date when the workflow stopped
          format: date-time
        status:
          allOf:
            - $ref: '#/components/schemas/WorkflowStatus'
          description: The status of the workflow; can be one of ['CANCELED', 'COMPLETED', 'FAILED', 'RUNNING', 'TERMINATED', 'TIMED_OUT']
        duration_seconds:
          title: Duration Seconds
          type: number
          description: The running time in seconds of the workflow
        parameters:
          title: Parameters
          type: object
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
    Page_Workflow_:
      title: Page[Workflow]
      required:
        - items
        - total
        - page
        - size
      type: object
      properties:
        items:
          title: Items
          type: array
          items:
            $ref: '#/components/schemas/Workflow'
        total:
          title: Total
          minimum: 0
          type: integer
        page:
          title: Page
          minimum: 1
          type: integer
        size:
          title: Size
          minimum: 1
          type: integer
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    MoleculeInput:
      title: MoleculeInput
      type: object
      properties:
        filename:
          title: Filename
          type: string
        base64data:
          title: Base64Data
          type: string
        id:
          title: Id
          type: string
          format: uuid4
        filetype:
          $ref: '#/components/schemas/ValidFileExtensions'
    ConformerSearchParams:
      title: ConformerSearchParams
      type: object
      properties:
        charge:
          title: Charge
          type: integer
          description: Set the charge of molecule
        multiplicity:
          title: Multiplicity
          type: integer
          description: Multiplicity (2S + 1) of the electronic state
        max_filter_jobs:
          title: Max Filter Jobs
          type: integer
          description: Maximum number of filter nodes for parallelization
        confgen:
          title: Confgen
          type: boolean
          description: Generate initial conformers using RDKits ETKG method
        confgen_max_n_conformers:
          title: Confgen Max N Conformers
          type: integer
          description: Number of initial conformers using RDKits ETKG method
        confgen_rmsd_threshold:
          title: Confgen Rmsd Threshold
          type: number
          description: RMSD threshold for conformer pruning after RDKits ETKG method
      additionalProperties: false
...
    System:
      title: System
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/SystemParams'
...
    JkBuilder:
      title: JkBuilder
      type: object
      properties:
        type:
          title: Type
          type: string
          default: core_dfjk
        params:
          $ref: '#/components/schemas/JkBuilderParams'
...
    XcBuilder:
      title: XcBuilder
      type: object
      properties:
        params:
          $ref: '#/components/schemas/XcBuilderParams'
...
    Est:
      title: Est
      type: object
      properties:
        params:
          $ref: '#/components/schemas/EstParams'
...
    ConformerSearchFilter:
      title: ConformerSearchFilter
      required:
        - filtertype
      type: object
      properties:
        filtertype:
          $ref: '#/components/schemas/FilterType'
        params:
          $ref: '#/components/schemas/ConformerSearchFilterParams'
        system:
          $ref: '#/components/schemas/System'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        est:
          $ref: '#/components/schemas/Est'
        hf:
          $ref: '#/components/schemas/FilterHf'
    ConformerSearchInitializationInputSpec:
      title: ConformerSearchInitializationInputSpec
      required:
        - molecule
        - params
        - filters
      type: object
      properties:
        molecule:
          $ref: '#/components/schemas/MoleculeInput'
        params:
          $ref: '#/components/schemas/ConformerSearchParams'
        filters:
          title: Filters
          type: array
          items:
            $ref: '#/components/schemas/ConformerSearchFilter'
    CreateConformerSearchWorkflowRequest:
      title: CreateConformerSearchWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - ConformerSearch
          type: string
          default: ConformerSearch
        parameters:
          $ref: '#/components/schemas/ConformerSearchInitializationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
...
    Hf:
      title: Hf
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/HfParams'
        outputs:
          $ref: '#/components/schemas/HfOutputs'
...
    Rcis:
      title: Rcis
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/RcisParams'
        outputs:
          $ref: '#/components/schemas/RcisOutputs'
...
    Pes:
      title: Pes
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/PesParams'
...
    LBFGSOptimization:
      title: LBFGSOptimization
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/LBFGSOptimizationParams'
        outputs:
          $ref: '#/components/schemas/LBFGSOptimizationOutputs'
    GeometryOptimizationInputSpec:
      title: GeometryOptimizationInputSpec
      required:
        - molecule
        - system
      type: object
      properties:
        molecule:
          $ref: '#/components/schemas/MoleculeInput'
        system:
          $ref: '#/components/schemas/System'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
        est:
          $ref: '#/components/schemas/Est'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        optimization:
          $ref: '#/components/schemas/LBFGSOptimization'
    CreateGeometryOptimizationWorkflowRequest:
      title: CreateGeometryOptimizationWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - GeometryOptimization
          type: string
          default: GeometryOptimization
        parameters:
          $ref: '#/components/schemas/GeometryOptimizationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
...
    MoleculeInputWithChargeAndMultiplicity:
      title: MoleculeInputWithChargeAndMultiplicity
      type: object
      properties:
        filename:
          title: Filename
          type: string
        base64data:
          title: Base64Data
          type: string
        id:
          title: Id
          type: string
          format: uuid4
        filetype:
          $ref: '#/components/schemas/ValidFileExtensions'
        params:
          $ref: '#/components/schemas/ChargeAndMultiplicity'
...
    InteractionEnergy:
      title: InteractionEnergy
      type: object
      properties:
        params:
          $ref: '#/components/schemas/InteractionEnergyParams'
    InteractionEnergyCalculationInputSpec:
      title: InteractionEnergyCalculationInputSpec
      required:
        - molecule_a
        - molecule_b
        - system
      type: object
      properties:
        molecule_a:
          $ref: '#/components/schemas/MoleculeInputWithChargeAndMultiplicity'
        molecule_b:
          $ref: '#/components/schemas/MoleculeInputWithChargeAndMultiplicity'
        system:
          $ref: '#/components/schemas/System'
        interaction_energy:
          $ref: '#/components/schemas/InteractionEnergy'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        est:
          $ref: '#/components/schemas/Est'
    CreateInteractionEnergyCalculationWorkflowRequest:
      title: CreateInteractionEnergyCalculationWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - InteractionEnergyCalculation
          type: string
          default: InteractionEnergyCalculation
        parameters:
          $ref: '#/components/schemas/InteractionEnergyCalculationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
...
    MoleculeInputWithGeometryOptimize:
      title: MoleculeInputWithGeometryOptimize
      type: object
      properties:
        filename:
          title: Filename
          type: string
        base64data:
          title: Base64Data
          type: string
        id:
          title: Id
          type: string
          format: uuid4
        filetype:
          $ref: '#/components/schemas/ValidFileExtensions'
        params:
          $ref: '#/components/schemas/GeometryOptimize'
        outputs:
          $ref: '#/components/schemas/MoleculeOutputs'
...
    InterpolatedPath:
      title: InterpolatedPath
      type: object
      properties:
        params:
          $ref: '#/components/schemas/InterpolatedPathParams'
...
    NebPath:
      title: NebPath
      type: object
      properties:
        params:
          $ref: '#/components/schemas/NebPathParams'
...
    FIREOptimizer:
      title: FIREOptimizer
      type: object
      properties:
        params:
          $ref: '#/components/schemas/FIREOptimizerParams'
    ReactionPathOptimizationInputSpec:
      title: ReactionPathOptimizationInputSpec
      required:
        - reactant
        - product
        - system
      type: object
      properties:
        reactant:
          $ref: '#/components/schemas/MoleculeInputWithGeometryOptimize'
        product:
          $ref: '#/components/schemas/MoleculeInputWithGeometryOptimize'
        system:
          $ref: '#/components/schemas/System'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
        optimization:
          $ref: '#/components/schemas/LBFGSOptimization'
        interpolation:
          $ref: '#/components/schemas/InterpolatedPath'
        neb:
          $ref: '#/components/schemas/NebPath'
        fire:
          $ref: '#/components/schemas/FIREOptimizer'
    CreateReactionPathOptimizationWorkflowRequest:
      title: CreateReactionPathOptimizationWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - ReactionPathOptimization
          type: string
          default: ReactionPathOptimization
        parameters:
          $ref: '#/components/schemas/ReactionPathOptimizationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
    SinglePointCalculationInputSpec:
      title: SinglePointCalculationInputSpec
      required:
        - molecule
        - system
        - hf
      type: object
      properties:
        molecule:
          $ref: '#/components/schemas/MoleculeInput'
        system:
          $ref: '#/components/schemas/System'
        hf:
          $ref: '#/components/schemas/Hf'
        est:
          $ref: '#/components/schemas/Est'
        rcis:
          $ref: '#/components/schemas/Rcis'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
    CreateSinglePointCalculationWorkflowRequest:
      title: CreateSinglePointCalculationWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - SinglePointCalculation
          type: string
          default: SinglePointCalculation
        parameters:
          $ref: '#/components/schemas/SinglePointCalculationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
...
    Torsion:
      title: Torsion
      required:
        - constraint
      type: object
      properties:
        constraint:
          $ref: '#/components/schemas/TorsionConstraint'
        interpolation:
          $ref: '#/components/schemas/RotationalPath'
    TorsionScanInitializationInputSpec:
      title: TorsionScanInitializationInputSpec
      required:
        - molecule
        - system
        - torsion
      type: object
      properties:
        molecule:
          $ref: '#/components/schemas/MoleculeInput'
        system:
          $ref: '#/components/schemas/System'
        torsion:
          $ref: '#/components/schemas/Torsion'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
        est:
          $ref: '#/components/schemas/Est'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        optimization:
          $ref: '#/components/schemas/LBFGSOptimization'
    CreateTorsionScanWorkflowRequest:
      title: CreateTorsionScanWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - TorsionScan
          type: string
          default: TorsionScan
        parameters:
          $ref: '#/components/schemas/TorsionScanInitializationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
    Preconditioner:
      title: Preconditioner
      required:
        - system
      type: object
      properties:
        system:
          $ref: '#/components/schemas/System'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        est:
          $ref: '#/components/schemas/Est'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
...
    TransitionStateOptimizationInputSpec:
      title: TransitionStateOptimizationInputSpec
      required:
        - molecule
        - system
      type: object
      properties:
        molecule:
          $ref: '#/components/schemas/MoleculeInput'
        system:
          $ref: '#/components/schemas/System'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        est:
          $ref: '#/components/schemas/Est'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
        preconditioner:
          $ref: '#/components/schemas/Preconditioner'
        optimization:
          $ref: '#/components/schemas/PRFOOptimizer'
    CreateTransitionStateOptimizationWorkflowRequest:
      title: CreateTransitionStateOptimizationWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - TransitionStateOptimization
          type: string
          default: TransitionStateOptimization
        parameters:
          $ref: '#/components/schemas/TransitionStateOptimizationInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
    TransitionStateOptimizationFromEndpointsInputSpec:
      title: TransitionStateOptimizationFromEndpointsInputSpec
      required:
        - reactant
        - product
        - system
      type: object
      properties:
        reactant:
          $ref: '#/components/schemas/MoleculeInputWithGeometryOptimize'
        product:
          $ref: '#/components/schemas/MoleculeInputWithGeometryOptimize'
        system:
          $ref: '#/components/schemas/System'
        jk_builder:
          $ref: '#/components/schemas/JkBuilder'
        xc_builder:
          $ref: '#/components/schemas/XcBuilder'
        est:
          $ref: '#/components/schemas/Est'
        hf:
          $ref: '#/components/schemas/Hf'
        rcis:
          $ref: '#/components/schemas/Rcis'
        pes:
          $ref: '#/components/schemas/Pes'
        preconditioner:
          $ref: '#/components/schemas/Preconditioner'
        optimization:
          $ref: '#/components/schemas/LBFGSOptimization'
        interpolation:
          $ref: '#/components/schemas/InterpolatedPath'
        neb:
          $ref: '#/components/schemas/NebPath'
        fire:
          $ref: '#/components/schemas/FIREOptimizer'
        prfo:
          $ref: '#/components/schemas/PRFOOptimizer'
    CreateTransitionStateOptimizationFromEndpointsWorkflowRequest:
      title: CreateTransitionStateOptimizationFromEndpointsWorkflowRequest
      required:
        - name
        - parameters
        - resources
      type: object
      properties:
        version:
          title: Version
          enum:
            - v1
          type: string
          default: v1
        name:
          title: Name
          type: string
          description: The user-defined name of the workflow
        kind:
          title: Kind
          enum:
            - TransitionStateOptimizationFromEndpoints
          type: string
          default: TransitionStateOptimizationFromEndpoints
        parameters:
          $ref: '#/components/schemas/TransitionStateOptimizationFromEndpointsInputSpec'
        resources:
          $ref: '#/components/schemas/ResourceRequest'
      additionalProperties: false
    WorkflowNumericResults:
      title: WorkflowNumericResults
      required:
        - id
        - kind
        - api_version
        - status
        - results
      type: object
      properties:
        id:
          title: Id
          type: string
          format: uuid4
        kind:
          title: Kind
          type: string
        api_version:
          title: Api Version
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        results:
          title: Results
          type: object
      additionalProperties: false
...